String & Enum support

+++ src/main/universal/res/jdk/java/lang/Math.java
@@ -2,15 +2,15 @@ package java.lang;
 
 class Math {
     /*@
+    // The standard doesn't have this one, but since int literals are not yet cast to doubles automatically, this one is in Math too.
     ghost
-    public static pure double sqrt(double x) {
+    public static pure int sqrt(int x) {
         // Blatantly wrong!
         return x;
     }
 
-    // The standard doesn't have this one, but since int literals are not yet cast to doubles automatically, this one is in Math too.
     ghost
-    public static pure int sqrt(int x) {
+    public static pure double sqrt(double x) {
         // Blatantly wrong!
         return x;
     }
